Move frames into subfolder
