+++ lib/frames/frame.d.ts
