+++ lib/frames/frame.js
