+++ lib/index.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 "use strict";
-var frame_1 = require("./frame");
+var frame_1 = require("./frames/frame");
 exports.Frame = frame_1.Frame;
 var clear = require("clear");
 var chalk = require("chalk");
